requirements.txt
asgiref==3.8.1
Django>=4.2
powerline-status==2.7
PyMySQL==1.1.1
pytz==2025.2
sqlparse==0.5.3
PyJWT>=2.9.0
